+++ lib/psql-cm.rb
@@ -3,7 +3,12 @@ require 'fileutils'
 require 'ostruct'
 require 'uri'
 require 'delegate'
+require 'tempfile'
 
 require_relative 'psql-cm/base'
 require_relative 'psql-cm/database'
+require_relative 'psql-cm/setup'
+require_relative 'psql-cm/dump'
+require_relative 'psql-cm/restore'
+require_relative 'psql-cm/submit'