+++ lib/psql-cm/base.rb
@@ -20,14 +20,13 @@ module PSQLCM
     def databases
       @databases = db.
         exec("SELECT datname as name FROM pg_database WHERE datname !~ 'template*|postgres';").
-        map {|row| row["name"]}
+        map {|row| row['name']}
 
       if config.databases.to_a.empty?
-        halt! "A list of databases must be given:\n  --databases={database_one}[,{database_two}[,...]]"
+        halt! 'A list of databases must be given:\n  --databases={database_one}[,{database_two}[,...]]'
       else # filter out databases not specified.
         @databases.select!{ |name| config.databases.include?(name) }
       end
-
       debug "databases> #{@databases}"
       @databases
     end
@@ -35,7 +34,10 @@ module PSQLCM
     def schemas(name = 'postgres')
       @schemas = db(name).
         exec("SELECT nspname AS name FROM pg_namespace WHERE nspname !~ '^pg_.*|information_schema';").
-        map{|row| row["name"]}
+        map{|row| row['name']}
+
+      # Filter out schemas not specified, if specified.
+      @schemas.select!{ |name| config.schemas.include?(name) } if config.schemas
       debug "schemas> #{@schemas}"
       @schemas
     end
@@ -44,149 +46,88 @@ module PSQLCM
       return @tree if @tree
       @tree = {}
       databases.each do |dbname|
-        @tree[dbname] = {}
-        schemas(dbname).each do |schema|
-          @tree[dbname][schema] = ['base.sql', 'cm.sql']
-        end
+        @tree[dbname] = schemas(dbname)
       end
-      debug "tree> tree: #{@tree}"
+      debug "tree> #{@tree}"
       @tree
     end
 
-    def dump!
-      unless config.sql_path
-        $stdout.puts "Warning: --sql-path was not set, defaulting to $PWD/sql."
-        config.sql_path = "#{ENV["PWD"]}/sql"
-      end
-
-      debug "dump> sql_path: #{config.sql_path}"
-      FileUtils.mkdir(config.sql_path) unless File.directory?(config.sql_path)
-      Dir.chdir(config.sql_path) do
-        tree.each_pair do |database, schema_hash|
-          debug "dump> database: #{database}"
-
-          File.directory?(File.join(config.sql_path,database)) or
-            FileUtils.mkdir(File.join(config.sql_path,database))
-
-          schema_hash.each_pair do |schema, files|
-          debug "dump> schema: #{schema}"
-            File.directory?(File.join(config.sql_path,database,schema)) or
-              FileUtils.mkdir(File.join(config.sql_path,database,schema))
-
-            base_file = File.join(config.sql_path,database,schema,'base.sql')
-            cm_file = File.join(config.sql_path,database,schema,'cm.sql')
-
-            FileUtils.touch(base_file)
-            FileUtils.touch(cm_file)
-
-            command = "pg_dump #{db(database).psql_args} --schema-only --no-owner --schema=#{schema} "
-            if File.size(base_file) > 0
-              command += "--file=#{cm_file}  --table=psql_cm "
-            else
-              command += "--file=#{base_file} --exclude-table=psql_cm "
-            end
-            command += "#{database}"
-            sh 'dump', command
-          end
-        end
-
-        sh 'git', "git init; git add ." unless File.exists?('.git') && File.directory?('.git')
-
-        sh 'git', "git commit -a -m 'PostgreSQL Change Management (psql-cm).\nDatabases: #{databases}\nTree: #{tree}'"
-      end
-    end # def dump!
-
-    def setup!
-      # Create psql_cm tables for each schema on the target db.
-      debug "setup> Setting up pg_psql_cm table in each target schema."
-      tree.each_pair do |database, schema_hash|
-        schema_hash.keys.each do |schema|
-          sql = <<-SQL
-            SET search_path = #{schema}, public;
-            CREATE TABLE IF NOT EXISTS #{schema}.pg_psql_cm
-            (
-              id bigserial NOT NULL PRIMARY KEY ,
-              is_base boolean NOT NULL,
-              created_at timestamp with time zone DEFAULT now(),
-              implementer text NOT NULL,
-              content text NOT NULL
-            );
-          SQL
-          debug "setup:#{database}:#{schema}> sql\n#{sql}"
-          db(database).exec sql
-        end
-      end
-    end
-
-    def restore!
-      unless config.sql_path
-        $stdout.puts "Warning: --sql-path was not set, defaulting to $PWD/sql."
-        config.sql_path = "#{ENV["PWD"]}/sql"
-      end
-
-      debug "restore> sql_path: #{config.sql_path}"
-      File.directory?(config.sql_path) or
-        halt! "Cannot restore from a path that does not exist: #{config.sql_path}"
-
-      Dir.chdir(config.sql_path) do
-
-        Dir['*'].each do |database|
-          next unless File.directory? database
-
-          Dir.chdir(database) do
-            sh "restore", "createdb #{db(database).psql_args} #{database}"
-
-            debug "restore:#{database}>"
-            Dir['*'].each do |schema|
-              next unless File.directory? schema
-
-              base_file = File.join(config.sql_path,database,schema,'base.sql')
-              cm_file = File.join(config.sql_path,database,schema,'cm.sql')
-
-              debug "restore:#{database}:#{schema}> #{base_file}"
-              sh 'restore', "psql #{db(database).psql_args} #{database} < #{base_file}"
-
-              next if File.size(cm_file) == 0
-
-              debug "restore:#{database}:#{schema}> #{cm_file}"
-              sh 'restore', "psql #{db(database).psql_args} #{database} < #{cm_file}"
-            end
-          end
-        end
-      end
-    end # def restore!
-
-    def change!
-      puts "TODO: allow change string and/or file to be specified and add to the
-      specified database scema control table"
-    end
-
-    def run!(action = config.action, parent_id = config.parent_id)
+    def run!(action = config.action)
       case action
-      when "console"
+      when 'console'
         require 'psql-cm/cli'
         ::PSQLCM::Console.run!
-      when "dump"
+      when 'dump'
         dump!
-      when "restore"
+      when 'restore'
         restore!
-      when "setup"
+      when 'setup'
         setup!
-      when "change"
-        change!
+      when 'submit'
+        submit!
       else
         halt! "An action must be given! {setup, dump, restore}" if action.nil?
         halt! "Action '#{action}' is not handled."
       end
     end
 
-    def sh(tag, command)
-      debug "sh:#{tag}> #{command}"
-      %x[#{command} 2>&1 | awk '! /already exists|create implicit/']
+    def sh(command)
+      debug "$ #{command}"
+      %x[#{command} 2>&1 | awk '! /NOTICE/']
+    end
+
+    def uri
+      return config.uri unless config.uri.to_s.empty?
+      $stdout.puts "NOTICE: uri is not set, defaulting to postgres://127.0.0.1:5432 (format: postgres://{user}:{password}@{host}:{port}/{database}\nwhere user, password, port and database are optional)"
+      config.uri = "postgres://127.0.0.1:5432"
+    end
+
+    private
+
+    def ensure_database_exists(database)
+      begin
+        db('postgres').exec("CREATE DATABASE #{database};")
+        debug "create> #{database}"
+      rescue => error
+        raise error unless error.message =~ /already exists/
+      end
+    end
+
+    def ensure_schema_exists(database,schema)
+      begin
+        db(database).exec("CREATE SCHEMA #{schema};")
+        debug "create> #{database}.#{schema}"
+      rescue => error
+        raise error unless error.message =~ /already exists/
+      end
+    end
+
+    def ensure_cm_table_exists(database,schema)
+      sql = <<-SQL
+        SET search_path = #{schema};
+        SET client_min_messages = warning;
+        CREATE TABLE IF NOT EXISTS #{schema}.#{config.cm_table}
+        (
+          id bigserial NOT NULL PRIMARY KEY ,
+          is_base boolean NOT NULL,
+          created_at timestamp with time zone DEFAULT now(),
+          implementer text NOT NULL,
+          content text NOT NULL
+        );
+      SQL
+      debug "create> #{database}.#{schema}.#{config.cm_table}"
+      db(database).exec(sql)
+    end
+
+    def sql_path
+      return config.sql_path unless config.sql_path.to_s.empty?
+      $stdout.puts "NOTICE: sql_path is not set, defaulting to #{ENV["PWD"]}/sql"
+      config.sql_path = "#{ENV["PWD"]}/sql"
     end
 
   end # class << self
 end
 
 ::PSQLCM.config.debug = !!ENV['DEBUG']
+::PSQLCM.config.cm_table = 'pg_psql_cm' # Default --cm-table name.