+++ lib/psql-cm/cli.rb
@@ -6,36 +6,40 @@ module PSQLCM
     class << self
       def parse!(arguments)
         ::OptionParser.new do |options|
-          options.banner = "Usage: psql-cm [options]"
-          options.separator ""
-          options.separator "Specific options:"
+          options.banner = 'Usage: psql-cm [options]'
+          options.separator ''
+          options.separator 'Specific options:'
 
-          options.on("-s", "--sql-path PATH", "Path to dump SQL cm files into.") do |path|
+          options.on('-s', '--sql-path PATH', 'Path to dump SQL cm files into.') do |path|
             ::PSQLCM.config.sql_path = path
           end
 
-          options.on("-a", "--databases NAMES", "A comma separated list of databases to cm.") do |names|
+          options.on('-a', '--databases=NAMES', 'A comma separated list of databases to cm.') do |names|
             ::PSQLCM.config.databases = names.split(',')
           end
 
-          options.on("-u", "--uri URI", "Path to the sink database connection file.") do |uri|
+          options.on('-u', '--uri=URI', 'Path to the sink database connection file.') do |uri|
             ::PSQLCM.config.uri = uri
           end
 
-          options.on("-D", "--[no-]debug", "Output debugging information.") do |debug|
+          options.on('-D', '--[no-]debug', 'Output debugging information.') do |debug|
             ::PSQLCM.config.debug = debug.nil? ? false : true
           end
 
-          options.on("-v", "--[no-]verbose", "Output verbosley.") do |verbose|
+          options.on('-v', '--[no-]verbose', 'Output verbosley.') do |verbose|
             ::PSQLCM.config.verbose = verbose
           end
 
-          options.on_tail("-h", "--help", "Print help and exit.") do
+          options.on('-c', '--cm-table=TABLE') do |cm_table|
+            ::PSQLCM.config.cm_table = cm_table
+          end
+
+          options.on_tail('-h', '--help', 'Print help and exit.') do
             puts options
             exit 0
           end
 
-          options.on_tail("--version", "Print version and exit.") do
+          options.on_tail('--version', 'Print version and exit.') do
             require 'psql-cm/version'
             puts ::PSQLCM::Version
             exit 0