+++ History.md
@@ -1,3 +1,15 @@
+# 0.0.6 - 2012-04-20
+
+Merged {base.sql, cm.sql} into a single sql-path/{database}/{schema}.sql file.
+
+Base is now the first change in the {schema}.sql CM file.
+
+Made Configuration Managment table name configurable via cli using -c|--cm-table=
+
+Made debug message output more structured and clear.
+
+Created {drop,create,setup,dump,restore} rake tasks for qa-ing.
+
 # 0.0.5 - 2012-04-18
 
 Added git repository feature to sql-path for dump action, a commit per run of